Move players logic to a specific module.
